fix: threat creation to move object
the threat calculation was fixed when it was on the board
even after a move, the original calculation stayed - verified this by
manually pulling a copy of the board's threat calculation, moving a piece
and then comparing `old_threats` to `board.threats`.  They should not
have been equal; however, they were.

Migrated threat creation to the Move object so that when a movement is
created, threats are then calculated, used and tossed - there's no
reason for this to stick around as it changes every time someone moves.