refactor: migrate en passant calculations to its own object
Pry en passant out of Board to thin out the Board class and
separate areas of concern.  En passant routes through move - called
every time a move is made to calculate whether or not it's an
en passant, and is set as an object on Board so that the previous move's
en passant state can be read.